# optimize str::from_utf8() validation when slice contains multibyte chars and str.chars().count() in all cases