… library/core/benches/str.rs