@the8472 2021-09-10
perf run since since `utf8_is_cont_byte` is also used in other places and I only benched `str.chars().count()`

@bors try @rust-timer queue