@rust-log-analyzer 2021-09-10
The job **`mingw-check`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/runs/3571796307) [(plain)](https://github.com/rust-lang/rust/commit/6df1a6337a0e6b3f98ae1d0fcb7b9804d28c7f6e/checks/3571796307/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
configure: rust.channel         := nightly
configure: rust.debug-assertions := True
configure: llvm.assertions      := True
configure: dist.missing-tools   := True
configure: build.configure-args := ['--enable-sccache', '--disable-manage-submodu ...
configure: writing `config.toml` in current directory
configure: 
configure: run `python /checkout/x.py --help`
configure: 
---
skip untracked path cpu-usage.csv during rustfmt invocations
skip untracked path src/doc/book/ during rustfmt invocations
skip untracked path src/doc/rust-by-example/ during rustfmt invocations
skip untracked path src/llvm-project/ during rustfmt invocations
Diff in /checkout/library/core/benches/str.rs at line 1:
-use test::{black_box, Bencher};
 use std::str;
+use test::{black_box, Bencher};
 
 const LOREM_SHORT: &str = "Lorem ipsum";
 
Running `"/checkout/obj/build/x86_64-unknown-linux-gnu/stage0/bin/rustfmt" "--config-path" "/checkout" "--edition" "2018" "--unstable-features" "--skip-children" "--check" "/checkout/library/core/tests/num/i128.rs" "/checkout/library/core/benches/num/flt2dec/strategy/dragon.rs" "/checkout/library/core/tests/num/uint_macros.rs" "/checkout/library/core/tests/num/mod.rs" "/checkout/library/core/benches/str.rs" "/checkout/library/core/tests/array.rs" "/checkout/library/core/tests/str.rs" "/checkout/library/core/benches/num/flt2dec/strategy/grisu.rs"` failed.
If you're running `tidy`, try again with `--bless`. Or, if you just want to format code, run `./x.py fmt` instead.
```

</details>