@the8472 2021-09-11
let's see if it was due to the extra function calls even though they should be inlined.

@bors try @rust-timer queue