@the8472 2021-09-11
Several of the improved and regressed benchmarks spend less/more time in `LLVM_passes`, `finish_ongoing_codegen` and `run_linker` so my guess is that this is mostly llvm noise rather than rust code being affected.